+++ Samples/iOS-Swift/iOS-Swift/ViewController.swift
@@ -7,6 +7,7 @@ class ViewController: UIViewController {
     @IBOutlet weak var anrFullyBlockingButton: UIButton!
     @IBOutlet weak var anrFillingRunLoopButton: UIButton!
     @IBOutlet weak var framesLabel: UILabel!
+    @IBOutlet weak var breadcrumbLabel: UILabel!
     
     private let dispatchQueue = DispatchQueue(label: "ViewController")
 
@@ -18,6 +19,7 @@ class ViewController: UIViewController {
             scope.setEnvironment("debug")
             scope.setTag(value: "swift", key: "language")
             scope.setExtra(value: String(describing: self), key: "currentViewController")
+
             let user = Sentry.User(userId: "1")
             user.email = "tony@example.com"
             scope.setUser(user)
@@ -28,8 +30,8 @@ class ViewController: UIViewController {
             if let data = "hello".data(using: .utf8) {
                 scope.add(Attachment(data: data, filename: "log.txt"))
             }
-            
         }
+
         // Also works
         let user = Sentry.User(userId: "1")
         user.email = "tony1@example.com"
@@ -43,7 +45,6 @@ class ViewController: UIViewController {
                 self.dsnTextField.backgroundColor = UIColor.systemGreen
             }
         }
-        
     }
     
     override func viewDidAppear(_ animated: Bool) {
@@ -53,6 +54,20 @@ class ViewController: UIViewController {
             Timer.scheduledTimer(withTimeInterval: 0.5, repeats: true) { _ in
                 self.framesLabel?.text = "Frames Total:\(PrivateSentrySDKOnly.currentScreenFrames.total) Slow:\(PrivateSentrySDKOnly.currentScreenFrames.slow) Frozen:\(PrivateSentrySDKOnly.currentScreenFrames.frozen)"
             }
+        }
+
+        SentrySDK.configureScope { (scope) in
+            let dict = scope.serialize()
+
+            guard
+                let crumbs = dict["breadcrumbs"] as? [[String: Any]],
+                let breadcrumb = crumbs.last,
+                let data = breadcrumb["data"] as? [String: String]
+            else {
+                return
+            }
+
+            self.breadcrumbLabel.text = "{ category: \(breadcrumb["category"] ?? "nil"), parentViewController: \(data["parentViewController"] ?? "nil"), beingPresented: \(data["beingPresented"] ?? "nil"), window_isKeyWindow: \(data["window_isKeyWindow"] ?? "nil"), is_window_rootViewController: \(data["is_window_rootViewController"] ?? "nil") }"
         }
     }